refactor/ streamline assembler/lifecycle.py and separate between block building input/context/config/output

+++ src/landseg/geopipe/foundation/data_blocks/assembler/io.py
@@ -66,7 +66,7 @@ class RasterReadOutput:
     image_padded_dem: numpy.ndarray | None
     image_nodata: float
     label_array: numpy.ndarray | None
-    label_nodata: float | None
+    label_nodata: int | None
 
 
 def check_npz_integrity(
